Encapsulate the s3_client
