+++ src/lib.rs
@@ -10,7 +10,7 @@ pub use self::web::Server;
 pub mod db;
 mod docbuilder;
 mod error;
-mod storage;
+pub mod storage;
 #[cfg(test)]
 mod test;
 pub mod utils;
